refactor: removing repeated code
copyfile and copyfile with replace moved to sinlge file just like rabbotai conversation
