+++ replacer/scripts/service-application/csproj.mjs
@@ -1,9 +1,6 @@
-import fs from 'fs';
+import { copyProjectFile } from '../../utils/file-operations.mjs';
 
-export default function (filePath, outputFilePath) {
-  fs.copyFileSync(filePath, outputFilePath);
-  return [];
-}
+export default copyProjectFile;
 
 export const file = 'src/services/academic/Looplex.DotNet.Samples.Academic.Application/Looplex.DotNet.Samples.Academic.Application.csproj';
 export const outputFile = '{PROJECT_PATH}/services/{{MODULE_NAME_CC}}/{{PROJECT_NAMESPACE}}.{MODULE_NAME}.Application/{{PROJECT_NAMESPACE}}.{MODULE_NAME}.Application.csproj';
