+++ replacer/scripts/service-domain/csproj.mjs
@@ -1,8 +1,6 @@
-import fs from 'fs';
+import { copyProjectFile } from '../../utils/file-operations.mjs';
 
-export default function (filePath, outputFilePath) {
-    fs.copyFileSync(filePath, outputFilePath);
-    return [];
-}
+export default copyProjectFile;
+
 export const file = 'src/services/academic/Looplex.DotNet.Samples.Academic.Domain/Looplex.DotNet.Samples.Academic.Domain.csproj';
 export const outputFile = '{PROJECT_PATH}/services/{{MODULE_NAME_CC}}/{{PROJECT_NAMESPACE}}.{MODULE_NAME}.Domain/{{PROJECT_NAMESPACE}}.{MODULE_NAME}.Domain.csproj';
