+++ replacer/scripts/webapi/dockerignore.mjs
@@ -1,9 +1,6 @@
-import fs from 'fs';
+import { copyProjectFile } from '../../utils/file-operations.mjs';
 
-export default function (filePath, outputFilePath) {
-    fs.copyFileSync(filePath, outputFilePath);
-    return [];
-}
+export default copyProjectFile;
 
 export const file = 'src/Looplex.DotNet.Samples.WebAPI/.dockerignore';
 export const outputFile = '{{WEB_API_PROJECT_NAME}}/.dockerignore';
