+++ replacer/scripts/webapi/policy.mjs
@@ -1,26 +1,17 @@
-import fs from 'fs';
-
-export default function (filePath, outputFilePath) {
-  // Read the file content
-  let content = fs.readFileSync(filePath, 'utf8');
-
-  const replaces = [
-    {
-      original: 'UserService',
-      find: /UserService/g,
-      replace: '{{SERVICE_NAME}}'
-    }
-  ]
-
-  replaces.forEach(replace => {
-    content = content
-      .replace(replace.find, replace.replace);
+import { CommandProcessor } from '../../utils/commandProcessor.mjs';
+export default (filePath, outputFilePath) => 
+  CommandProcessor.process({
+    filePath,
+    outputFilePath,
+    patterns: replaces
   });
-  fs.writeFileSync(outputFilePath, content, 'utf8');
-  console.log(`Processed file saved to: ${outputFilePath}`);
-
-  return replaces;
-}
 
 export const file = 'src/Looplex.DotNet.Samples.WebAPI/policy.csv';
 export const outputFile = '{{WEB_API_PROJECT_NAME}}/policy.csv';
+const replaces = [
+  {
+    original: 'UserService',
+    find: /UserService/g,
+    replace: '{{SERVICE_NAME}}'
+  }
+]
